pyproject.toml
[project]
name = "vegan-bulletin"
version = "0.1.0"
description = "A web application to display an aggregated set of news and events related to veganism"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "pillow>=10.0.0",
    "python-decouple>=3.8",
    "django-recurrence>=1.11.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
